Add affinity to base class so it is inherited by every experiment
