+++ experiments/saxpy/experiment.py
@@ -10,8 +10,6 @@ from benchpark.openmp import OpenMPExperiment
 from benchpark.cuda import CudaExperiment
 from benchpark.rocm import ROCmExperiment
 from benchpark.caliper import Caliper
-from benchpark.affinity import Affinity
-
 
 class Saxpy(
     Experiment,
@@ -19,7 +17,6 @@ class Saxpy(
     CudaExperiment,
     ROCmExperiment,
     Caliper,
-    Affinity,
 ):
     variant(
         "workload",
